backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
gunicorn>=21.2.0
Pillow>=10.3.0
django-cors-headers>=4.3.1
djoser>=2.2.0